+++ CHANGELOG.md
@@ -12,6 +12,7 @@ All notable changes to `bol-retailer-api` will be documented in this file.
 - Added Products -> Get a list of competing offers by EAN
 - Added Products -> Get Product Placement
 - Added Products -> Get Product Ids By Eancode
+- Added Products -> Get Product Ratings By Eancode
 
 ### Changes
 - Drop php7 support